+++ include/ntgdi.h
@@ -364,7 +364,7 @@ BOOL     WINAPI NtGdiPlgBlt( HDC hdc, const POINT *point, HDC hdc_src, INT x_src
                              INT width, INT height, HBITMAP mask, INT x_mask, INT y_mask,
                              DWORD bk_color );
 BOOL     WINAPI NtGdiPolyDraw(HDC hdc, const POINT *points, const BYTE *types, DWORD count );
-ULONG    WINAPI NtGdiPolyPolyDraw( HDC hdc, const POINT *points, const UINT *counts,
+ULONG    WINAPI NtGdiPolyPolyDraw( HDC hdc, const POINT *points, const ULONG *counts,
                                    DWORD count, UINT function );
 BOOL     WINAPI NtGdiPtInRegion( HRGN hrgn, INT x, INT y );
 BOOL     WINAPI NtGdiPtVisible( HDC hdc, INT x, INT y );
